# Conflicts:
#	common/pom.xml
#	contrib/data/pom.xml
#	contrib/data/tpch-sample-data/pom.xml
#	contrib/format-maprdb/pom.xml
#	contrib/format-maprdb/src/main/java/org/apache/drill/exec/store/mapr/db/MapRDBSubScan.java
#	contrib/gis/pom.xml
#	contrib/native/client/src/protobuf/UserBitShared.pb.cc
#	contrib/native/client/src/protobuf/UserBitShared.pb.h
#	contrib/pom.xml
#	contrib/sqlline/pom.xml
#	contrib/storage-hbase/pom.xml
#	contrib/storage-hive/core/pom.xml
#	contrib/storage-hive/hive-exec-shade/pom.xml
#	contrib/storage-hive/pom.xml
#	contrib/storage-jdbc/pom.xml
#	contrib/storage-kafka/pom.xml
#	contrib/storage-kudu/pom.xml
#	contrib/storage-mongo/pom.xml
#	contrib/storage-opentsdb/pom.xml
#	distribution/pom.xml
#	drill-yarn/pom.xml
#	exec/java-exec/pom.xml
#	exec/java-exec/src/main/java/org/apache/drill/exec/store/dfs/easy/EasyWriter.java
#	exec/jdbc-all/pom.xml
#	exec/jdbc/pom.xml
#	exec/memory/base/pom.xml
#	exec/memory/pom.xml
#	exec/pom.xml
#	exec/rpc/pom.xml
#	exec/vector/pom.xml
#	logical/pom.xml
#	pom.xml
#	protocol/pom.xml
#	protocol/src/main/java/org/apache/drill/exec/proto/UserBitShared.java
#	protocol/src/main/java/org/apache/drill/exec/proto/beans/CoreOperatorType.java
#	protocol/src/main/protobuf/UserBitShared.proto
#	tools/fmpp/pom.xml
#	tools/pom.xml

+++ exec/java-exec/src/main/java/org/apache/drill/exec/planner/sql/DrillOperatorTable.java
@@ -17,9 +17,9 @@
  */
 package org.apache.drill.exec.planner.sql;
 
-import com.google.common.collect.ArrayListMultimap;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
+import org.apache.drill.shaded.guava.com.google.common.collect.ArrayListMultimap;
+import org.apache.drill.shaded.guava.com.google.common.collect.Lists;
+import org.apache.drill.shaded.guava.com.google.common.collect.Maps;
 import org.apache.calcite.sql.SqlAggFunction;
 import org.apache.calcite.sql.SqlFunction;
 import org.apache.calcite.sql.fun.SqlBetweenOperator;
@@ -54,7 +54,7 @@ public class DrillOperatorTable extends SqlStdOperatorTable {
   private final ArrayListMultimap<String, SqlOperator> drillOperatorsWithInferenceMap = ArrayListMultimap.create();
   // indicates remote function registry version based on which drill operator were loaded
   // is used to define if we need to reload operator table in case remote function registry version has changed
-  private long functionRegistryVersion;
+  private int functionRegistryVersion;
 
   private final OptionManager systemOptionManager;
 
@@ -70,14 +70,14 @@ public class DrillOperatorTable extends SqlStdOperatorTable {
    *
    * @param version registry version
    */
-  public void setFunctionRegistryVersion(long version) {
+  public void setFunctionRegistryVersion(int version) {
     functionRegistryVersion = version;
   }
 
   /**
    * @return function registry version based on which operator table was loaded
    */
-  public long getFunctionRegistryVersion() {
+  public int getFunctionRegistryVersion() {
     return functionRegistryVersion;
   }
